chore: move storage validator tests to their own file
